30/01/25

+++ printf/ft_printf.c
@@ -6,15 +6,15 @@
 /*   By: robenite <robenite@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/26 21:31:53 by robenite          #+#    #+#             */
-/*   Updated: 2025/01/20 21:45:27 by robenite         ###   ########.fr       */
+/*   Updated: 2025/01/30 12:21:08 by robenite         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
 static int	format_chek(char format, va_list args);
-static int	pointer_f(unsigned int args, char *base);
-static int	hexa_f(unsigned int args, char *base);
+static int	pointer_f(unsigned long args, char *base);
+static int	hexa_f(unsigned long args, char *base);
 static int	ft_putnbr_ish_no_sign_fd(unsigned long args);
 
 int	ft_printf(char const *text, ...)
@@ -51,24 +51,24 @@ static int	format_chek(char format, va_list args)
 	else if (format == 's')
 		r_counter = ft_putstr_ish_fd(va_arg(args, char *), 1);
 	else if (format == 'p')
-		r_counter = pointer_f(va_arg(args, unsigned int),
+		r_counter = pointer_f(va_arg(args, unsigned long),
 				"0123456789abcdef");
 	else if (format == 'd' || format == 'i')
 		r_counter = ft_putnbr_ish_fd(va_arg(args, int), 1);
 	else if (format == 'u')
 		r_counter = ft_putnbr_ish_no_sign_fd(va_arg(args, unsigned long));
 	else if (format == 'x')
-		r_counter = hexa_f(va_arg(args, unsigned int),
+		r_counter = hexa_f(va_arg(args, unsigned long),
 				"0123456789abcdef");
 	else if (format == 'X')
-		r_counter = hexa_f(va_arg(args, unsigned int),
+		r_counter = hexa_f(va_arg(args, unsigned long),
 				"0123456789ABCDEF");
 	else if (format == '%')
 		r_counter = ft_putchar_ish_fd('%', 1);
 	return (r_counter);
 }
 
-static int	pointer_f(unsigned int args, char *base)
+static int	pointer_f(unsigned long args, char *base)
 {
 	int	sum;
 
@@ -83,7 +83,7 @@ static int	pointer_f(unsigned int args, char *base)
 	return (sum);
 }
 
-static int	hexa_f(unsigned int args, char *base)
+static int	hexa_f(unsigned long args, char *base)
 {
 	int	sum;
